[refactor] 탭 이벤트 처리 로그 수정 및 커스텀 예외 사용하도록 수정

+++ backend/src/main/java/coffeeshout/cardgame/domain/CardGameScore.java
@@ -33,7 +33,7 @@ public class CardGameScore extends MiniGameScore {
 
     @Override
     public long getValue() {
-        return addition * multiplier;
+        return (long) addition * multiplier;
     }
 }
